1. Split server to routes and model. Changed hardcoded roots on client. Stuck on WS proxy
